refactor: separate filter config file
