+++ src/components/MusicManager/useFilterState.ts
@@ -1,92 +1,6 @@
-import { O, pipe } from "@mobily/ts-belt";
 import { useCallback, useMemo, useReducer } from "react";
 
-import {
-  type FetchTracksOptions,
-  type SortField,
-  type SortOrder,
-} from "@/lib/api/tracks";
-
-type NonNullableProps<T> = { [K in keyof T]-?: NonNullable<T[K]> };
-export type FilterState = NonNullableProps<FetchTracksOptions>;
-
-export const PAGE_SIZES = [10, 20, 30, 50, 100];
-const DEFAULT_PAGE_SIZE = 20;
-
-const SORT_FIELD_LABELS: Record<SortField, string> = {
-  createdAt: "Newest",
-  title: "Title",
-  artist: "Artist",
-  album: "Album",
-} as const;
-
-const SORT_ORDER_LABELS: Record<SortOrder, string> = {
-  asc: "Ascending",
-  desc: "Descending",
-} as const;
-
-export const SORT_OPTIONS = Object.entries(SORT_FIELD_LABELS).map(
-  ([value, label]) => ({ value, label })
-);
-
-export const SORT_ORDER_OPTIONS = Object.entries(SORT_ORDER_LABELS).map(
-  ([value, label]) => ({ value, label })
-);
-
-const isSortField = (value: unknown): value is SortField =>
-  typeof value === "string" && value in SORT_FIELD_LABELS;
-
-const isSortOrder = (value: unknown): value is SortOrder =>
-  typeof value === "string" && value in SORT_ORDER_LABELS;
-
-const safeParseInt = (v: string): O.Option<number> =>
-  O.fromPredicate(parseInt(v, 10), (n) => !isNaN(n));
-
-export type FilterConfig = {
-  readonly [K in keyof FilterState]: {
-    readonly defaultValue: FilterState[K];
-    readonly parse: (value: string) => O.Option<FilterState[K]>;
-  };
-};
-
-export const filterConfig: FilterConfig = {
-  search: {
-    defaultValue: "",
-    parse: (v) => O.Some(v.trim()),
-  },
-  genre: {
-    defaultValue: "",
-    parse: (v) => O.Some(v.trim()),
-  },
-  sort: {
-    defaultValue: "createdAt",
-    parse: (v) => (isSortField(v) ? O.Some(v) : O.None),
-  },
-  order: {
-    defaultValue: "desc",
-    parse: (v) => (isSortOrder(v) ? O.Some(v) : O.None),
-  },
-  page: {
-    defaultValue: 1,
-    parse: (v) =>
-      pipe(
-        safeParseInt(v),
-        O.filter((n) => n > 0)
-      ),
-  },
-  limit: {
-    defaultValue: DEFAULT_PAGE_SIZE,
-    parse: (v) =>
-      pipe(
-        safeParseInt(v),
-        O.filter((n) => PAGE_SIZES.includes(n))
-      ),
-  },
-  artist: {
-    defaultValue: "",
-    parse: (v) => O.Some(v.trim()),
-  },
-};
+import { type FilterState, filterConfig } from "./filter.config";
 
 type FilterAction =
   | {
